uint16 for serial sizes; reset if recv full

+++ src/onp/channel-serial.h
@@ -3,8 +3,8 @@
 
 typedef void (*channel_serial_connect_cb)(char*);
 
-void channel_serial_init(channel_serial_connect_cb cb);
-int  channel_serial_recv(char* b, int l);
-int  channel_serial_send(char* b, int n);
+void     channel_serial_init(channel_serial_connect_cb cb);
+uint16_t channel_serial_recv(char* b, uint16_t l);
+uint16_t channel_serial_send(char* b, uint16_t n);
 
 #endif
